Adicionando funcionalidade de pesquisa
