+++ src/App.tsx
@@ -1,19 +1,17 @@
-import {useState} from 'react';
+
 import {Global} from './styles/global'
 import {PeoplesProvider} from './hooks/useSearch'
 //Components
 import{Header} from './components/Header/index'
-import {Search} from './components/Search/index'
-import { DataTable } from './components/DataTable';
+import { Table } from './components/Table';
 
 function App() {
 
   return (
     <PeoplesProvider>
-      <Global/>
-      <Header/>
-      <Search/>
-      <DataTable/>
+        <Global/>
+        <Header/>
+        <Table/>
     </PeoplesProvider>
   );
 }
