--- src/components/Search/style.ts
@@ -1,30 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.div`
-  height: 7.15vw;
-  width:100vw;
-
-  display:flex;
-  flex-direction:row;
-  justify-content: space-between;
-  margin-bottom:4vw ;
-
-
-  p{
-    font-size:1.6vw;
-    margin: 6.9vw 0 0 5vw;
-    font-weight:500;
-  }
-
-
-  input[type="text"]{
-    margin: 6.9vw 5vw 0 0;
-    height:2.7vw;
-    width:19.5vw;
-    border-radius:0.4vw;
-
-    padding:0.5vw 0.8vw;
-    border: 1px solid var(--border-color);
-    background: var(--card-color);
-  }
-`;
