+++ src/components/Table/style.ts
@@ -1,23 +1,48 @@
 import styled from 'styled-components'
 
-
 export const Container = styled.div`
-  height: 38.15vw;
-  width:90.6vw;
-  background:var(--background);
-  margin: 6.9vw 5vw 0 5vw;
-
-  text-transform:uppercase;
-  font-size:1.1vw;
+  width: 100vw;
+  display:flex;
+  flex-direction:column;
+  justify-content:space-between;
 
 
-  div{
+  .Search{
+    height: 7.15vw;
+    font-size:1.6vw;
+
+    display:flex;
+    flex-direction:row;
+    justify-content:space-between;
+
+    padding: 2vw 5vw 0 5vw;
+
+    input[type="search"]{
+
+    height:2.7vw;
+    width:19.5vw;
+    border-radius:0.4vw;
+
+    padding:0.5vw 0.8vw;
+    border: 1px solid var(--border-color);
+    background: var(--card-color);
+  }
+
+  input[type="search"]:focus{
+    background:rgba(255,255,255,0.05);
+  }
+  }
+
+
+  .Table-header{
     background:var(--header-table);
-    padding:0.80% 2%;
-
+    padding:12px 15px 12px 2%;
     display: grid;
     grid-template-rows: 25px;
     grid-template-columns: 0.5fr repeat(4, 2fr);
+    margin:0 75px;
+
+    font-size:1.2vw;
 
     color:var(--card-color);
     .--icon{
@@ -42,19 +67,21 @@ export const Container = styled.div`
       text-align:center;
     }
   }
+  }
 
-`;
-
-export const Card = styled.main`
+  .Table-item{
     display: grid;
     grid-template-rows: 39px;
     grid-row-gap: 2px;
     grid-template-columns: 0.5fr repeat(4, 2fr);
-    
+
     padding:0.80% 2%;
     background:var(--card-color);
+    margin:0 75px;
 
-   
+    font-size:1.2vw;
+
+
     color:#000;
     .--item-icon{
       text-align:center;
@@ -81,5 +108,7 @@ export const Card = styled.main`
       height:2.29vw;
       width:34px;
     }
-    
+
+  }
 `;
+
