+++ src/hooks/useSearch.tsx
@@ -1,7 +1,6 @@
 import{createContext,useState,useEffect, useContext} from 'react';
 import { api } from '../services/api';
 
-
 interface peoplesProps {
   id: number;
   name: string;
@@ -16,9 +15,11 @@ interface PeoplesProviderProps{
 }
 interface PeoplesContextData {
   peoples:peoplesProps [];
+  setPeoples:any;
+  PhoneFormat:any;
 }
 
-type peoplesInput = Omit<peoplesProps, 'id' | 'admission_date' | 'phone' | 'image'>; 
+
  const PeoplesContext = createContext<PeoplesContextData>(
   {} as PeoplesContextData
   );
@@ -26,13 +27,31 @@ type peoplesInput = Omit<peoplesProps, 'id' | 'admission_date' | 'phone' | 'imag
 export function PeoplesProvider({children} : PeoplesProviderProps){
   const [peoples, setPeoples] = useState<peoplesProps[]>([]);
 
+  const PhoneFormat = (phone:string) =>{
+    let array = phone.split('');
+    array.splice(9,0, '-');
+    array.splice(0,0, '+');
+    array.splice(3,0,' ');
+    array.splice(4,0,'(');
+    array.splice(7,0,') ');
+
+    array.join('')
+    return array;
+  }
+
   useEffect(()=>{
     api.get('/employess')
     .then(res=> setPeoples(res.data))
     console.log(peoples)
   },[])
+
   return (
-    <PeoplesContext.Provider value={{peoples}}>
+    <PeoplesContext.Provider
+    value={{
+      peoples,
+      setPeoples,
+      PhoneFormat
+      }}>
       {children}
     </PeoplesContext.Provider>
   )
@@ -40,5 +59,5 @@ export function PeoplesProvider({children} : PeoplesProviderProps){
 export function useSearch(){
   const context = useContext(PeoplesContext)
 
-  return context 
+  return context
 }
